refactor: provide cpp file for utility functions
