+++ CMakeLists.txt
@@ -39,6 +39,7 @@ list(APPEND sources
 	"src/Config.cpp"
 	"src/Chain.cpp"
 	"src/Alias.cpp"
+	"src/Utility.cpp"
 	
 	"src/Console.cpp"
 	"src/Blizzard.cpp"
